+++ actions/add/source/index.js
@@ -4,7 +4,8 @@ const fs = require('zerodep/node/fs');
 const exec = require('zerodep/node/exec');
 // argument must be an url
 const url = process.argv[2];
-const pathArgs = fs.filePathResolveArgs({ parser: 'json' }, process.env.sourcePath, ...fs.pathResolveArgsFromUri(url, { pathOnly: false, subdomainsNested: true }));
+const sourcePath = process.argv[3] || process.env.sourcePath;
+const pathArgs = fs.filePathResolveArgs({ parser: 'json' }, sourcePath, ...fs.pathResolveArgsFromUri(url, { pathOnly: false, subdomainsNested: true }));
 // run async
 (async () => {
     if (!fs.isFile(...pathArgs)) {
@@ -16,11 +17,16 @@ const pathArgs = fs.filePathResolveArgs({ parser: 'json' }, process.env.sourcePa
             console.log(`Source url '${url}' successfully added.`);
             let exitCode = await exec('node', 'actions', 'add', 'schema', fs.pathResolve(...pathArgs.slice(0, -1)));
             if (exitCode) throw new Error(`Compiling the schema from source url '${url}' failed!`);
+            // in vscode if json.schemaDownload.enable = true the missing refs are automatically saved elsewhere
             const missingSources = require(`${process.env.PWD}/base/schema/sources/missing`)(...pathArgs.slice(0, -1));
-            missingSources.forEach(async (url) => {
-                exitCode = await exec('node', 'actions', 'add', 'source', url);
-                if (exitCode) throw new Error(`Compiling the schema from source url '${url}' failed!`);
-            });
+            // a missing source may embed other missing sources
+            if (missingSources.length) {
+                let ref = missingSources[0];
+                ref = ref.endsWith('/') ? ref + 'index.json' : !fs.pathExtName(ref) ? ref + '.json' : ref;
+                console.log(`Start adding schema dependency from source url: '${ref}'...`);
+                exitCode = await exec('node', 'actions', 'add', 'source', ref, sourcePath);
+                if (exitCode) throw new Error(`Adding the schema dependency from source url '${ref}' failed!`);
+            };
         } catch (err) {
             fs.removeFile(fs.pathResolve(...pathArgs));
             if (!fs.entries(...pathArgs.slice(0, -1)).length) fs.removeDir(...pathArgs.slice(0, -1));
@@ -28,7 +34,7 @@ const pathArgs = fs.filePathResolveArgs({ parser: 'json' }, process.env.sourcePa
             process.exit(1);
         }
     } else {
-        console.error(`Source url '${url}' already exists. If you want to renew first remove it from source then add it again!`);
+        console.error(`Source url '${url}' already exists. If you want to renew first remove it from source path then add it again!`);
         process.exit(1);
     }
 })();