+++ actions/remove/source/index.js
@@ -4,7 +4,8 @@ const fs = require('zerodep/node/fs');
 const exec = require('zerodep/node/exec');
 // argument must be an url
 const url = process.argv[2];
-const pathArgs = fs.filePathResolveArgs({ parser: 'json' }, process.env.sourcePath, ...fs.pathResolveArgsFromUri(url, { pathOnly: false, subdomainsNested: true }));
+const sourcePath = process.argv[3] || process.env.sourcePath;
+const pathArgs = fs.filePathResolveArgs({ parser: 'json' }, sourcePath, ...fs.pathResolveArgsFromUri(url, { pathOnly: false, subdomainsNested: true }));
 // run async
 (async () => {
     if (fs.isFile(...pathArgs)) {