+++ base/schema/sources/missing/index.js
@@ -10,8 +10,8 @@ module.exports = (...dirPathResolveArgs) => {
         (...refs) => {
             const target = fn.get(schema, ...refs);
             if (typeof target.$ref === 'string') {
-                const keys = fn.jsonPointerKeys(target.$ref.replace('#/', ''));
-                if (!Object.keys(schema).find((id) => fn.get(schema, id, ...keys.slice(1)))) sources.push(target.$ref);
+                const keys = fn.jsonPointerKeys(target.$ref.startsWith('#/') ? target.$ref.replace('#/', '') : target.$ref);
+                if (!Object.keys(schema).find((id) => fn.get(schema, id, ...keys.slice(1)))) sources.push(new URL('', keys.join('/').replace('/#/', '#/')).href);
             }
         },
         schema